Fix health connect settings not being deletable (#685)
* Add SettingsGroup superclass to all settings groups

Reimplements IntervalStoreManager via generator

* Call reset on all settings types

fixes hc settings not resettet

* Use new SettingsGroup type to save duplicated code

* Fix cross import
