+++ app/lib/model/storage/db/config_dao.dart
@@ -1,3 +1,5 @@
+import 'dart:collection';
+
 import 'package:blood_pressure_app/model/storage/db/config_db.dart';
 import 'package:blood_pressure_app/model/storage/db/settings_loader.dart';
 import 'package:blood_pressure_app/model/storage/export_columns_store.dart';
@@ -6,8 +8,10 @@ import 'package:blood_pressure_app/model/storage/export_pdf_settings.dart';
 import 'package:blood_pressure_app/model/storage/export_settings.dart';
 import 'package:blood_pressure_app/model/storage/export_xls_settings.dart';
 import 'package:blood_pressure_app/model/storage/health_connect_settings.dart';
-import 'package:blood_pressure_app/model/storage/interval_store.dart';
+import 'package:blood_pressure_app/model/storage/interval_store_manager.dart';
 import 'package:blood_pressure_app/model/storage/settings.dart';
+import 'package:blood_pressure_app/model/storage/types/interval_storage_setting.dart';
+import 'package:settings_annotation/settings_annotation.dart';
 
 /// Class for loading data from the database.
 ///
@@ -137,9 +141,9 @@ class ConfigDao implements SettingsLoader {
   @override
   Future<IntervalStoreManager> loadIntervalStorageManager() async {
     _intervallStorageInstance ??= IntervalStoreManager(
-      await _loadStore(0),
-      await _loadStore(1),
-      await _loadStore(2),
+      mainPage: await _loadStore(0),
+      exportPage: await _loadStore(1),
+      statsPage: await _loadStore(2),
     );
     return _intervallStorageInstance!;
   }
@@ -196,4 +200,8 @@ class ConfigDao implements SettingsLoader {
   @override
   Future<HealthConnectSettings> loadHealthConnectSettings() async =>
       HealthConnectSettings();
+
+  @override
+  UnmodifiableListView<SettingsGroup> get initializedSettings =>
+      throw UnimplementedError('Code unmaintained');
 }
