+++ app/lib/model/storage/db/file_settings_loader.dart
@@ -1,4 +1,4 @@
-import 'dart:convert';
+import 'dart:collection';
 import 'dart:io';
 
 import 'package:archive/archive_io.dart';
@@ -9,10 +9,10 @@ import 'package:blood_pressure_app/model/storage/export_pdf_settings.dart';
 import 'package:blood_pressure_app/model/storage/export_settings.dart';
 import 'package:blood_pressure_app/model/storage/export_xls_settings.dart';
 import 'package:blood_pressure_app/model/storage/health_connect_settings.dart';
-import 'package:blood_pressure_app/model/storage/interval_store.dart';
+import 'package:blood_pressure_app/model/storage/interval_store_manager.dart';
 import 'package:blood_pressure_app/model/storage/settings.dart';
-import 'package:flutter/widgets.dart';
 import 'package:path/path.dart';
+import 'package:settings_annotation/settings_annotation.dart';
 import 'package:sqflite/sqflite.dart';
 
 /// Store settings in a directory format on disk.
@@ -32,12 +32,12 @@ class FileSettingsLoader implements SettingsLoader {
   /// Instantiates a settings file relative to [_path] and writes changes.
   ///
   /// If the is read successfully [build] is called else [createNew] is called.
-  T _loadFile<T extends ChangeNotifier>(
+  T _loadFile<T extends SettingsGroup>(
     String fileName,
     T Function(String) build,
     T Function() createNew,
-    String Function(T) serialize,
   ) {
+    if (_instances.containsKey(fileName)) return _instances[fileName] as T;
     final f = File(join(_path, fileName));
     T? obj;
     try {
@@ -47,17 +47,22 @@ class FileSettingsLoader implements SettingsLoader {
     }
     obj ??= createNew();
 
-    obj.addListener(() => f.writeAsStringSync(serialize(obj!)));
-    f.writeAsStringSync(serialize(obj));
+    obj.addListener(() => f.writeAsStringSync(obj!.toJson()));
+    f.writeAsStringSync(obj.toJson());
+    _instances[fileName] = obj;
     return obj;
   }
+
+  final _instances = <String, SettingsGroup>{};
+
+  @override
+  UnmodifiableListView<SettingsGroup> get initializedSettings => UnmodifiableListView(_instances.values);
 
   @override
   Future<CsvExportSettings> loadCsvExportSettings() async => _loadFile(
     'csv-export',
     CsvExportSettings.fromJson,
     CsvExportSettings.new,
-    (e) => e.toJson(),
   );
 
   @override
@@ -65,7 +70,6 @@ class FileSettingsLoader implements SettingsLoader {
     'export-columns',
     ExportColumnsManager.fromJson,
     ExportColumnsManager.new,
-    (e) => e.toJson(),
   );
 
   @override
@@ -73,32 +77,13 @@ class FileSettingsLoader implements SettingsLoader {
     'export',
     ExportSettings.fromJson,
     ExportSettings.new,
-    (e) => e.toJson(),
   );
 
   @override
   Future<IntervalStoreManager> loadIntervalStorageManager() async => _loadFile(
     'intervall-store',
-    (String jsonStr) {
-      final json = jsonDecode(jsonStr);
-      if (json is Map<String, dynamic>) {
-        return IntervalStoreManager(
-          json['main'] is! String ? IntervalStorage()
-              : IntervalStorage.fromJson(json['main']! as String),
-          json['export'] is! String ? IntervalStorage()
-              : IntervalStorage.fromJson(json['export']! as String),
-          json['stats'] is! String ? IntervalStorage()
-              : IntervalStorage.fromJson(json['stats']! as String),
-        );
-      }
-      return IntervalStoreManager(IntervalStorage(), IntervalStorage(), IntervalStorage());
-    },
-    () => IntervalStoreManager(IntervalStorage(), IntervalStorage(), IntervalStorage()),
-    (e) => jsonEncode({
-      'main': e.mainPage.toJson(),
-      'export': e.exportPage.toJson(),
-      'stats': e.statsPage.toJson(),
-    }),
+    IntervalStoreManager.fromJson,
+    IntervalStoreManager.new,
   );
 
   @override
@@ -106,7 +91,6 @@ class FileSettingsLoader implements SettingsLoader {
     'pdf-export',
     PdfExportSettings.fromJson,
     PdfExportSettings.new,
-    (e) => e.toJson(),
   );
 
   @override
@@ -114,7 +98,6 @@ class FileSettingsLoader implements SettingsLoader {
     'xsl-export', // wrong spelling kept for compatability reasons
     ExcelExportSettings.fromJson,
     ExcelExportSettings.new,
-    (e) => e.toJson(),
   );
 
   @override
@@ -122,7 +105,6 @@ class FileSettingsLoader implements SettingsLoader {
     'general',
     Settings.fromJson,
     Settings.new,
-    (e) => e.toJson(),
   );
 
   @override
@@ -130,7 +112,6 @@ class FileSettingsLoader implements SettingsLoader {
     'health_connect',
     HealthConnectSettings.fromJson,
     HealthConnectSettings.new,
-    (e) => e.toJson(),
   );
 
   /// Attempt to backup all stored data to archive.
