+++ app/test/model/storage/db/file_settings_loader_test.dart
@@ -3,8 +3,8 @@ import 'dart:io';
 
 import 'package:blood_pressure_app/model/export_import/column.dart';
 import 'package:blood_pressure_app/model/storage/db/file_settings_loader.dart';
-import 'package:blood_pressure_app/model/storage/interval_store.dart';
 import 'package:blood_pressure_app/model/storage/types/export_format_setting.dart';
+import 'package:blood_pressure_app/model/storage/types/time_step.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
