+++ settings_annotation/lib/settings_annotation.dart
@@ -1,3 +1,4 @@
 export 'annotations.dart';
 export 'src/deep_setting.dart';
 export 'src/setting.dart';
+export 'src/settings_group.dart';
